+++ apps/web_demo/cli.yaml
@@ -8,6 +8,7 @@ meta:
     - "basic.terminal"
 
 target: "chat.terminal"
+hide_logs: true
 
 lifecycle:
   auto_start: true
